Added experience gaining and more configs
Added configs for experience gaining as well as actually giving the
experience points in the game events.

Added an option for starting heroes into configs, which lets you choose
which heroes are given when a player connects for the first time.

Commented herowars.py file so it should be a little clearer now than it
was before.