Fixed eargs format and added hero status msg
Skills now take eargs as keyword arguments instead of a dict, so use
on_event(self, **eargs) instead of on_event(self, eargs)
The benefit being the ability of taking some eargs arguments as a
parameter already in the definition:
def on_attack(self, defender, **eargs):
So you can choose what you wish to use and the rest are thrown in the
eargs dict.

Also added a message and translation for hero's status message which
gets shown on player_spawn.