Restrictions Update
Add hooks, events and lists needed to create a restriction system for
heroes. Just need to add a weapon name to the set and it will stop a
player from obtaining that weapon in anyway possible. Remove the weapon
from the set and they can obtain it again.

Example:
player.restrictions.add('weapon_ak47')